# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/resumes": {
      "get": {
        "operationId": "ResumesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeListDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "resumes"
        ]
      },
      "post": {
        "operationId": "ResumesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "resumes"
        ]
      }
    },
    "/api/resumes/{id}": {
      "get": {
        "operationId": "ResumesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "resumes"
        ]
      },
      "put": {
        "operationId": "ResumesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "resumes"
        ]
      },
      "delete": {
        "operationId": "ResumesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "resumes"
        ]
      }
    },
...
  },
  "info": {
    "title": "job-research API",
    "description": "REST API локального ассистента поиска работы",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "requestId": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "formErrors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "OkResponseDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
...
      "ResumeListDto": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number"
            },
            "userId": {
              "type": "number"
            },
            "title": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "skills": {
              "type": "string",
              "minLength": 1
            },
            "experienceSummary": {
              "type": "string",
              "minLength": 1
            },
            "experienceYears": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": false,
              "maximum": 80,
              "exclusiveMaximum": false
            },
            "education": {
              "type": "string",
              "nullable": true
            },
            "desiredSalary": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "nullable": true
            },
            "desiredSalaryCurrency": {
              "type": "string",
              "minLength": 1,
              "maxLength": 16,
              "nullable": true
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            }
          },
          "required": [
            "id",
            "userId",
            "title",
            "skills",
            "experienceSummary",
            "experienceYears",
            "education",
            "desiredSalary",
            "desiredSalaryCurrency",
            "createdAt",
            "updatedAt"
          ]
        }
      },
      "ResumeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "userId": {
            "type": "number"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "skills": {
            "type": "string",
            "minLength": 1
          },
          "experienceSummary": {
            "type": "string",
            "minLength": 1
          },
          "experienceYears": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 80,
            "exclusiveMaximum": false
          },
          "education": {
            "type": "string",
            "nullable": true
          },
          "desiredSalary": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "nullable": true
          },
          "desiredSalaryCurrency": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "skills",
          "experienceSummary",
          "experienceYears",
          "education",
          "desiredSalary",
          "desiredSalaryCurrency",
          "createdAt",
          "updatedAt"
        ]
      },
      "ResumeInputDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "skills": {
            "type": "string",
            "minLength": 1
          },
          "experienceSummary": {
            "type": "string",
            "minLength": 1
          },
          "experienceYears": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 80,
            "exclusiveMaximum": false
          },
          "education": {
            "type": "string",
            "nullable": true
          },
          "desiredSalary": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "nullable": true
          },
          "desiredSalaryCurrency": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16,
            "nullable": true
          }
        },
        "required": [
          "title",
          "skills",
          "experienceSummary",
          "experienceYears",
          "education",
          "desiredSalary",
          "desiredSalaryCurrency"
        ]
      },
...
    }
  }
}